+++ pkg/networkserver/mac/adr_param_setup_test.go
@@ -24,6 +24,7 @@ import (
 	. "go.thethings.network/lorawan-stack/v3/pkg/networkserver/internal"
 	. "go.thethings.network/lorawan-stack/v3/pkg/networkserver/internal/test"
 	. "go.thethings.network/lorawan-stack/v3/pkg/networkserver/mac"
+	"go.thethings.network/lorawan-stack/v3/pkg/specification/macspec"
 	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
 	"go.thethings.network/lorawan-stack/v3/pkg/util/test"
 	"go.thethings.network/lorawan-stack/v3/pkg/util/test/assertions/should"
@@ -164,7 +165,7 @@ func TestNeedsADRParamSetupReq(t *testing.T) {
 							},
 						},
 						Band:  phy,
-						Needs: conf.Needs && macVersion.Compare(ttnpb.MACVersion_MAC_V1_1) >= 0,
+						Needs: conf.Needs && macspec.UseADRParamSetupReq(macVersion),
 					},
 				)
 			})