+++ CHANGELOG.md
@@ -14,6 +14,7 @@ For details about compatibility between different releases, see the **Commitment
 - Allow setting an expiry date for API keys in the Console
 - New event type `gs.gateway.connection.stats` with connection statistics. A new event is sent at most every `gs.update-connection-stats-debounce-time` time and at least every `gs.update-connection-stats-interval` time.
 - Button to export as JSON end device `mac-settings` and `mac-state` in the Console.
+- Support for the `FOpts encryption, usage of FCntDwn` LoRaWAN 1.1 erratum.
 
 ### Changed